separate out evend handler
