+++ src/lib/fragments/ExtendableGraph.events.js
@@ -1,0 +1,62 @@
+import {contains, filter, has, isNil, type} from 'ramda';
+
+export const filterEventData = (gd, eventData, event) => {
+    let filteredEventData;
+    if (contains(event, ['click', 'hover', 'selected'])) {
+        const points = [];
+
+        if (isNil(eventData)) {
+            return null;
+        }
+
+        /*
+         * remove `data`, `layout`, `xaxis`, etc
+         * objects from the event data since they're so big
+         * and cause JSON stringify ciricular structure errors.
+         *
+         * also, pull down the `customdata` point from the data array
+         * into the event object
+         */
+        const data = gd.data;
+
+        for (let i = 0; i < eventData.points.length; i++) {
+            const fullPoint = eventData.points[i];
+            const pointData = filter(function (o) {
+                return !contains(type(o), ['Object', 'Array']);
+            }, fullPoint);
+            if (
+                has('curveNumber', fullPoint) &&
+                has('pointNumber', fullPoint) &&
+                has('customdata', data[pointData.curveNumber])
+            ) {
+                pointData.customdata =
+                    data[pointData.curveNumber].customdata[
+                        fullPoint.pointNumber
+                    ];
+            }
+
+            // specific to histogram. see https://github.com/plotly/plotly.js/pull/2113/
+            if (has('pointNumbers', fullPoint)) {
+                pointData.pointNumbers = fullPoint.pointNumbers;
+            }
+
+            points[i] = pointData;
+        }
+        filteredEventData = {points};
+    } else if (event === 'relayout' || event === 'restyle') {
+        /*
+         * relayout shouldn't include any big objects
+         * it will usually just contain the ranges of the axes like
+         * "xaxis.range[0]": 0.7715822247381828,
+         * "xaxis.range[1]": 3.0095292008680063`
+         */
+        filteredEventData = eventData;
+    }
+    if (has('range', eventData)) {
+        filteredEventData.range = eventData.range;
+    }
+    if (has('lassoPoints', eventData)) {
+        filteredEventData.lassoPoints = eventData.lassoPoints;
+    }
+    return filteredEventData;
+};
